--- generators/Classes/TestClassNameTransformer.cs
@@ -1,9 +1,0 @@
-using Humanizer;
-
-namespace Generators.Classes
-{
-    public class TestClassNameTransformer : IStringTransformer
-    {
-        public string Transform(string input) => $"{input.Dehumanize()}Test";
-    }
-}
