--- generators/Classes/TestedClassNameTransformer.cs
@@ -1,9 +1,0 @@
-using Humanizer;
-
-namespace Generators.Classes
-{
-    public class TestedClassNameTransformer : IStringTransformer
-    {
-        public string Transform(string input) => input.Dehumanize();
-    }
-}
