--- generators/Helpers/To.cs
@@ -1,13 +1,0 @@
-﻿using Generators.Classes;
-using Generators.Methods;
-
-namespace Generators.Helpers
-{
-    public static class To
-    {
-        public static readonly TestedClassNameTransformer TestedClassName = new TestedClassNameTransformer();
-        public static readonly TestClassNameTransformer TestClassName = new TestClassNameTransformer();
-        public static readonly TestedMethodNameTransformer TestedMethodName = new TestedMethodNameTransformer();
-        public static readonly TestMethodNameTransformer TestMethodName = new TestMethodNameTransformer();
-    }
-}
