+++ generators/Input/CanonicalDataCaseJsonConverter.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Generators.Data
+namespace Generators.Input
 {
     public class CanonicalDataCaseJsonConverter : JsonConverter
     {
