+++ generators/Input/CanonicalDataCasesJsonConverter.cs
@@ -4,7 +4,7 @@ using System.Reflection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Generators.Data
+namespace Generators.Input
 {
     public class CanonicalDataCasesJsonConverter : JsonConverter
     {
