--- generators/Methods/TestedMethodNameTransformer.cs
@@ -1,9 +1,0 @@
-using Humanizer;
-
-namespace Generators.Methods
-{
-    public class TestedMethodNameTransformer : IStringTransformer
-    {
-        public string Transform(string input) => input.Dehumanize();
-    }
-}
