+++ generators/Output/EqualityTestMethodGenerator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
 
-namespace Generators.Methods
+namespace Generators.Output
 {
     public class EqualityTestMethodGenerator : TestMethodGenerator
     {
