+++ generators/Output/ExceptionTestMethodGenerator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Generators.Methods
+namespace Generators.Output
 {
     public class ExceptionTestMethodGenerator : TestMethodGenerator
     {
